fix(build): remove disk race from SPA entry resolvers, verify post-build
resolveSpaBundle/resolveEntryAssets gated their fixed-filename return on
fs.existsSync(dist/assets/<file>) at closeBundle time, on the theory that
plain Rollup assets (unlike dist/index.html) are reliably present by then.
False for this build: Rollup's write phase doesn't reliably finish within
any closeBundle hook, and resolveSpaBundle's hard throw on that race took
down every locale's deploy for 20+ hours (0/60 successful runs since
2026-07-25T17:30, confirmed pre-existing to and unrelated to #4760's
content — the old poll-based code failed identically).

Both resolvers now return the fixed filenames unconditionally, no disk
I/O — callers only need the string, and Rollup guarantees the file exists
by the time the vite build process exits. Actual on-disk presence is
verified once, after the process exits, by the new
scripts/verify-spa-entry-assets.mjs, chained after every vite build entry
point in package.json.

Sibling-pattern audit (AGENTS.md #6) found one real but lower-severity
match: shared/chunkFiles.ts's 3 callers do the same closeBundle-time
fs.readdirSync(dist/assets), but degrade gracefully (missing chunk = no
preload, never a build failure) rather than crash/corrupt — tracked
separately in #4762 rather than folded into this urgent hotfix.
