+++ docs/AGENTS-HISTORY.md
@@ -128,3 +128,11 @@ Issue #4745 (e i 3 predecessori #4638/#4738/#4746) erano lo stesso deploy failur
 Eliminati come conseguenza: `build-plugins/shared/spaBundleRx.ts` (regex condivise, non più importate da nessuno), `spaBundleResolverPrewarmPlugin` (il plugin di prewarm-ordering, reso inutile), l'env var CI `SPA_BUNDLE_POLL_TIMEOUT_MS` (rimossa da 6 workflow — non più letta da nessun codice).
 
 Lezione: quando un fix ricorrente (3+ volte sullo stesso sintomo) continua a ri-fallire nonostante tuning progressivo dello stesso parametro (timeout, ordine), è il segnale che il meccanismo stesso va messo in discussione — non il valore del parametro. Verificare prima se la premessa originale del meccanismo (qui: "il filename è scopribile solo a runtime") è ancora vera nel codice attuale.
+
+**Follow-up 2026-07-26 — la premessa "plain asset file affidabile a closeBundle-time" era falsa**: subito dopo il merge del fix sopra (`c42dc566ef8`), il deploy è rimasto rosso su tutti e 4 i locale — ma un controllo su `gh run list` ha mostrato che ANCHE il commit immediatamente precedente al merge (`a91b93df049`, ancora sul vecchio codice poll-based) falliva con lo stesso sintomo di fondo: `poll exhausted after 2392 attempts (120000 ms). Last state: index.html does not exist yet.` — con `cpu_s=0.24` su `wall_s=120.01`, cioè il poll ha aspettato inerte 120s pieni senza che Rollup finisse mai di scrivere `dist/index.html`. Verifica su `gh run list --branch main --workflow deploy.yml --limit 60`: **zero successi** in una finestra di ~20 ore (`2026-07-25T17:30` → `2026-07-26T13:14`), sito live fermo su un build di `2026-07-24T06:34` (>2 giorni stale). Il deploy era quindi già rosso PRIMA del merge del fix precedente, per lo stesso identico motivo — non una regressione introdotta da quella PR.
+
+**Causa radice reale (seconda iterazione)**: il fix precedente aveva sostituito il poll su `dist/index.html` con un singolo `fs.existsSync` non-retrying su `dist/assets/<file>`, sulla premessa che i plain asset file (a differenza di `dist/index.html`) fossero affidabilmente presenti a `closeBundle`-time — premessa mutuata dal precedente `shared/chunkFiles.ts`. Falsa: per QUESTO build (corpus grande, migliaia di chunk), la fase di scrittura-su-disco di Rollup non finisce in modo affidabile entro NESSUN hook `closeBundle`, né per `dist/index.html` né per i plain asset. `resolveSpaBundle` (throw-on-missing, usato da `ogPagesPlugin.ts`) buttava giù la build istantaneamente; `resolveEntryAssets` (degrado silenzioso a `''`, usato da 125 simboli via GitNexus) avrebbe silenziosamente spedito pagine SEO senza script di hydration — stessa classe di bug già documentata nell'incidente 2026-04-30 (`scripts/audit-spa-bundle-injection.mjs`).
+
+**Fix strutturale (secondo giro)**: eliminato il disk-check del tutto da `resolveSpaBundle`/`resolveEntryAssets` — tornano i due filename fissi come costanti pure, senza mai leggere `dist/assets/`. Corretto perché i chiamanti hanno bisogno solo della STRINGA del filename per generare l'HTML, non dell'esistenza fisica del file al momento della generazione — Rollup garantisce che il file esiste per quando il processo `vite build` termina, molto dopo che qualunque pagina che lo referenzia è stata scritta. La verifica di esistenza fisica si sposta in un passo POST-build separato, `scripts/verify-spa-entry-assets.mjs` (eseguito con `tsx`, stesso pattern di `scripts/generate-health-facilities-jobs.mjs`), incatenato dopo `vite build` in tutti gli script `build`/`build:ci`/`build:prod`/`build:fast`/`prepush:fast` — a quel punto Node ha già atteso che ogni scrittura pendente si flush-i, quindi non c'è più nessuna race da rincorrere.
+
+Lezione (bis): un "precedente di successo" (`chunkFiles.ts` mai fallito con lo stesso pattern) non è una prova che il pattern sia corretto in astratto — può solo significare che nessuno ha ancora osservato il caso limite (qui: corpus abbastanza grande da spingere la fase di scrittura oltre qualunque finestra `closeBundle`). Quando un check di esistenza-su-disco serve solo a validare un valore già noto a monte (qui: il filename, fisso da `vite.config.ts`), la soluzione più robusta è NON fare il check nel punto critico (mid-build, in gara con l'I/O) ma spostarlo a un punto dove la garanzia è strutturale (dopo che il processo dell'intero build è terminato), invece di continuare a tunare il meccanismo di check stesso.
