+++ package.json
@@ -8,9 +8,9 @@
     "dev": "vite",
     "predev": "node scripts/assemble-jobs-dataset.mjs --stats",
     "prebuild": "node scripts/generate-image-thumbnails.mjs && node scripts/prebuild-webp.mjs && node scripts/assemble-jobs-dataset.mjs --stats",
-    "build": "NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build",
-    "build:ci": "NODE_OPTIONS='--max-old-space-size=14336 --expose-gc' vite build --minify esbuild",
-    "build:prod": "NODE_OPTIONS='--max-old-space-size=18432' vite build",
+    "build": "NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build && npx tsx scripts/verify-spa-entry-assets.mjs",
+    "build:ci": "NODE_OPTIONS='--max-old-space-size=14336 --expose-gc' vite build --minify esbuild && npx tsx scripts/verify-spa-entry-assets.mjs",
+    "build:prod": "NODE_OPTIONS='--max-old-space-size=18432' vite build && npx tsx scripts/verify-spa-entry-assets.mjs",
     "preview": "vite preview",
     "test": "vitest run",
     "test:fast": "vitest run --bail 1",
@@ -20,8 +20,8 @@
     "test:e2e:visual": "playwright test tests/e2e/seo-visual-regression.spec.ts",
     "test:e2e:visual:update": "playwright test tests/e2e/seo-visual-regression.spec.ts --update-snapshots",
     "build:e2e": "npm run build && npm run test:e2e:smoke",
-    "build:fast": "SKIP_FUEL_DAILY=1 SKIP_WEEKLY_EMPLOYERS=1 SKIP_JOB_MARKET_SNAPSHOT=1 SKIP_HEALTH_PREMIUMS=1 SKIP_ORPHAN_LANDINGS=1 SKIP_BORDER_WAIT=1 NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build",
-    "prepush:fast": "SKIP_FUEL_DAILY=1 SKIP_WEEKLY_EMPLOYERS=1 SKIP_JOB_MARKET_SNAPSHOT=1 SKIP_HEALTH_PREMIUMS=1 SKIP_ORPHAN_LANDINGS=1 SKIP_BORDER_WAIT=1 node scripts/assemble-jobs-dataset.mjs --stats && vitest run --bail 1 && NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build",
+    "build:fast": "SKIP_FUEL_DAILY=1 SKIP_WEEKLY_EMPLOYERS=1 SKIP_JOB_MARKET_SNAPSHOT=1 SKIP_HEALTH_PREMIUMS=1 SKIP_ORPHAN_LANDINGS=1 SKIP_BORDER_WAIT=1 NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build && npx tsx scripts/verify-spa-entry-assets.mjs",
+    "prepush:fast": "SKIP_FUEL_DAILY=1 SKIP_WEEKLY_EMPLOYERS=1 SKIP_JOB_MARKET_SNAPSHOT=1 SKIP_HEALTH_PREMIUMS=1 SKIP_ORPHAN_LANDINGS=1 SKIP_BORDER_WAIT=1 node scripts/assemble-jobs-dataset.mjs --stats && vitest run --bail 1 && NODE_OPTIONS='--max-old-space-size=18432' FAST_BUILD=1 vite build && npx tsx scripts/verify-spa-entry-assets.mjs",
     "test:articles": "vitest run tests/article-seo-fallback.test.ts tests/static-pages-blog-skip.test.ts tests/article-duplicate-detection.test.ts tests/article-tax-content-guard.test.ts tests/related-articles.test.ts tests/whats-new.test.tsx",
     "test:seo-gates": "vitest run tests/title-length.test.ts tests/h1-not-equal-title.test.ts tests/url-max-length.test.ts tests/hreflang-no-broken.test.ts tests/dist-single-h1-per-page.test.ts tests/dist-link-anchor-text.test.ts tests/seo-html-lang-sync.test.ts tests/seo/software-application-jsonld.test.ts",
     "gate:seo-source": "vitest run tests/seo/ --reporter=dot",
